+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Committing with empty X-Nakadi-StreamId causes 503
+- Massive topic deletion while switching timelines is now made with small interval between deletions
 
 ### Added
 - Create event type with SAFE rack-awareness